changed package structure in eva-accession-deprecate module
