+++ eva-accession-deprecate/src/main/java/uk/ac/ebi/eva/accession/deprecate/batch/io/DeprecableClusteredVariantsReader.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.deprecate.io;
+package uk.ac.ebi.eva.accession.deprecate.batch.io;
 
 import com.mongodb.BasicDBObject;
 import com.mongodb.MongoClient;
