+++ eva-accession-deprecate/src/main/java/uk/ac/ebi/eva/accession/deprecate/batch/listeners/DeprecationStepProgressListener.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.deprecate.listeners;
+package uk.ac.ebi.eva.accession.deprecate.batch.listeners;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
