+++ eva-accession-deprecate/src/main/java/uk/ac/ebi/eva/accession/deprecate/configuration/BeanNames.java
@@ -17,14 +17,14 @@ package uk.ac.ebi.eva.accession.deprecate.configuration;
 
 public class BeanNames {
 
-    static final String DEPRECABLE_CLUSTERED_VARIANTS_READER = "DEPRECABLE_CLUSTERED_VARIANTS_READER";
+    public static final String DEPRECABLE_CLUSTERED_VARIANTS_READER = "DEPRECABLE_CLUSTERED_VARIANTS_READER";
 
-    static final String DEPRECATION_WRITER = "DEPRECATION_WRITER";
+    public static final String DEPRECATION_WRITER = "DEPRECATION_WRITER";
 
-    static final String DEPRECATE_CLUSTERED_VARIANTS_STEP = "DEPRECATE_CLUSTERED_VARIANTS_STEP";
+    public static final String DEPRECATE_CLUSTERED_VARIANTS_STEP = "DEPRECATE_CLUSTERED_VARIANTS_STEP";
 
-    static final String DEPRECATE_CLUSTERED_VARIANTS_JOB = "DEPRECATE_CLUSTERED_VARIANTS_JOB";
+    public static final String DEPRECATE_CLUSTERED_VARIANTS_JOB = "DEPRECATE_CLUSTERED_VARIANTS_JOB";
 
-    static final String DEPRECATION_PROGRESS_LISTENER = "DEPRECATION_PROGRESS_LISTENER";
+    public static final String DEPRECATION_PROGRESS_LISTENER = "DEPRECATION_PROGRESS_LISTENER";
 
 }
