+++ eva-accession-deprecate/src/main/java/uk/ac/ebi/eva/accession/deprecate/configuration/batch/io/DeprecationWriterConfiguration.java
@@ -13,22 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.deprecate.configuration;
+package uk.ac.ebi.eva.accession.deprecate.configuration.batch.io;
 
-import org.springframework.batch.core.listener.StepListenerSupport;
+import org.springframework.batch.core.configuration.annotation.StepScope;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.data.mongodb.core.MongoTemplate;
 
-import uk.ac.ebi.eva.accession.core.model.dbsnp.DbsnpClusteredVariantEntity;
-import uk.ac.ebi.eva.accession.deprecate.listeners.DeprecationStepProgressListener;
+import uk.ac.ebi.eva.accession.core.configuration.nonhuman.MongoConfiguration;
+import uk.ac.ebi.eva.accession.deprecate.batch.io.DeprecationWriter;
 
-import static uk.ac.ebi.eva.accession.deprecate.configuration.BeanNames.DEPRECATION_PROGRESS_LISTENER;
+import static uk.ac.ebi.eva.accession.deprecate.configuration.BeanNames.DEPRECATION_WRITER;
 
 @Configuration
-public class ListenerConfiguration {
+@Import({MongoConfiguration.class})
+public class DeprecationWriterConfiguration {
 
-    @Bean(DEPRECATION_PROGRESS_LISTENER)
-    public StepListenerSupport<DbsnpClusteredVariantEntity, DbsnpClusteredVariantEntity> deprecationProgressListener() {
-        return new DeprecationStepProgressListener();
+    @Bean(DEPRECATION_WRITER)
+    @StepScope
+    DeprecationWriter deprecationWriter(MongoTemplate mongoTemplate) {
+        return new DeprecationWriter(mongoTemplate);
     }
 }
