+++ eva-accession-deprecate/src/main/java/uk/ac/ebi/eva/accession/deprecate/configuration/batch/policies/ChunkSizeCompletionPolicyConfiguration.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.eva.accession.deprecate.configuration;
+package uk.ac.ebi.eva.accession.deprecate.configuration.batch.policies;
 
 import org.springframework.batch.core.configuration.annotation.StepScope;
 import org.springframework.batch.repeat.policy.SimpleCompletionPolicy;
