+++ eva-accession-deprecate/src/test/java/uk/ac/ebi/eva/accession/deprecate/batch/io/DeprecableClusteredVariantsReaderTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.deprecate.io;
+package uk.ac.ebi.eva.accession.deprecate.batch.io;
 
 import com.lordofthejars.nosqlunit.annotation.UsingDataSet;
 import com.lordofthejars.nosqlunit.mongodb.MongoDbConfigurationBuilder;
